refactor(api): use jpql for retrieving filtered articles

+++ src/main/java/ir/ac/ut/ece/ie/articles/Page.java
@@ -9,5 +9,5 @@ import java.util.List;
 @AllArgsConstructor
 public class Page<T> {
     List<T> content;
-    Integer total;
+    long total;
 }
